requirements.txt
jupyterlab
black
pytest
kafka-python
rich
pyspark
